Backend/requirements.txt
alembic==1.14.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RBAC==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
pipdeptree==2.23.4
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
setuptools==75.3.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.2
wheel==0.44.0
